pyproject.toml
[project]
name = "backend-ai-resume-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "bcrypt==4.0.1",
    "boto3>=1.42.31",
    "fastapi[standard]>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
